Fix a database migration issue with certain versions of MySQL.
MySQL's handling of default values for longtext fields is strict, and
are problematic during an evolution. We can't have a default value (it's
okay for it to be empty in certain versions), unless it's NULL. We
didn't have it set to allow NULL, though.

Now the new Repository.extra_data field allows NULL values, and is NULL
by default, restoring compatibility.

An evolution has been added for anyone who has successfully upgraded
already, or installed 1.6.7 on a new database.