Fix the Account field default for hosting services.
A change in the comparison for the Account field in the Repository form
broke the default for the Account field. Opening an existing repository
would show the Link New Account entry again. The proper repository could
be selected again, so this isn't fatal, but it's annoying.

This was caused because a "==" was turned into a "===" after review,
since === is ideal, but we were comparing a string and an integer.

This change makes this explicit by calling parseInt() on the selected
value in order to normalize the types.