Support checking if re-authorization is needed.
If a hosting service account needs re-authorization (as determined by
the is_authorized method, which can check the local account auth data for
known keys), we will now show this on the repository configuration page
with some helpful text.

Nothing committed needs this yet, but an in-progress Bitbucket change
will.