Pass the list of diffsets and public reviews to ReviewRequest.get_last_activity.
ReviewRequest.get_last_activity can now take a list of diffsets and
review to operate on. This works as an optimization when you don't need
that function to perform its own queries. We use this in review_detail
now in order to reduce the number of queries we need to make.