Reduce queries for reviews, replies, and comments.
Each reply to a review was generating queries for each comment, each
replied review, and each user. These ballooned up in size quite a bit.

We now cache these along with everything else we're caching to reduce
lookups. The lists of replies are all computed at one place when we're
computing everything else, and new accessors on Review and BaseComment
make it easy to grab the results without assuming a query.