Added live updates to timestamps.
Our timestamps have always shown that something happened "x minutes ago"
or "x years ago" or whatever. Unfortunately, with the caching, this
can get "stuck," and it's always wrong anyway since you may be on a page
for a while and the timestamps never updated.

Now we have live updating of all the timestamps. This works similarly to
Django's timesince, except it will update every minute. It also makes it
easier to cache the pages without having that data stored.

This makes use of a jQuery plugin I wrote, jquery-timesince, which
will be maintained at https://github.com/chipx86/jquery-timesince. This
plugin implements the same logic as Django's own timesince filter.