Add a framework for handling file attachments by mimetype.
This introduces a new framework for mimetype handlers. These handlers
can render file attachments for display in the review request. Right
now, we have handlers for generic file attachments (which don't render
anything), for text files (which renders the first couple lines of the
file), and for images (which render the image sort of like a
screenshot).

There are some things that still need to be done. For example, an
initially uploaded file attachment won't have any sort of thumbnail
rendered until the page is loaded. This will need to happen before
1.7.0.

Reviewed at http://reviews.reviewboard.org/r/2967/