Don't grind to a halt when encountering files with very long lines.
Files with very long lines would cause Review Board to sit processing
for a very, very long time. This happened both when styling the
orig/modified files using Pygments (which operates per-line, and is
really slow when there's just one giant line), and when attempting to
highlight changes between replaced lines.

Now we determine if there's a line over our limit (set to 1000
characters). If so, we turn off pygments, and we don't highlight changes
for that line (if it's a "replace").

This helps a great deal when people are putting up minified
JavaScript/CSS or large single-line data sets up for review.