Fix issues with spurious update bubbles and incorrect HTTP status codes.
The update bubbles on 1.7 have been broken for a while. Changing issue
statuses would cause an update bubble to appear, due to the timestamp
coming in not being in the correct format. We now set it exactly as the
Last Update resource sets it, by returning
last_activity_time.isoformat().

The Last Update resource was broken too. Every other check was returning
a 200 OK instead of a 304 Not Modified, when each one should be a 304
until there's actually new updates to report. This was due to the wrong
value being set in the Last Modified header, since we converted the
timestamp to ISO8601 format too soon. We now set the correct timestamp.