Fix extra Last Update check timers firing after changing issue statuses.
Every time we resolved or reopened an issue, we'd register again for
updates to the review request. However, we already had a check going, and
each new one caused a new check. This could mean we'd hammer the server
eventually.

We now have a RB.ReviewRequest.markUpdated function that sets the
last updated timestamp. We use this instead when resolving an issue. It
ensures that we don't interfere with the existing checks, but we do mark
that we know when the last activity was.