Fix close messages with the new optimizations.
In the old code, we were modifying data in the ChangeDescription's info,
when we were normalizing for display. This was a hack and only worked
because the latest_changedesc we grabbed was a separate copy from any of
the ChangeDescriptions we were modifying.

In the new optimized code, that's no longer the case. We're operating on
the same copy. This caused the status string comparisons to fail, since
we changed the names to human-readable strings. Now we make a deep copy
of the data we intend to modify so that doesn't happen.