Added "Show changes" links to interdiffs in changedesc blobs.
When a user uploads a diff, we show a link to the uploaded diff revision. We
should also provide a linke to the interdiff, since that's what a lot of people
will want to see.

This is based on Curtis miller's change http://reviews.reviewboard.org/r/2810/
I've updated it to fix the remaining issues, and tweaked the text a little bit
("Show changes" instead of "Show changes").

Testing done:
Published a few versions of a diff and verified that the new link showed
what I expected to.

Reviewed at http://reviews.reviewboard.org/r/3225/