Fix the order of displayed comments on a review.
Diff comments and replies were displayed in insert order rather than
timestamp order. This made things very confusing when comments were
created in orders other than top-down, or when replies were published in
orders other than the order created.

This was due to the ordering for diff comments not including the
timestamp, and instead taking into account only the filediffs and line
numbers. The timestamp is now factored in.

Fixes bug #2684