Properly handle draft and inactive screenshots/file attachments.
The recent optimizations caused a regression for screenshots and file
attachments where any draft entries wouldn't be shown for the user after
uploading until the review request was published. This was due to
querying the review request and not the draft for these.

There was also a performance issue where inactive screenshots/file
attachments weren't being cached, causing a lookup per comment
referencing an inactive one. Now we cache these, but only when we
encounter a case where we need them.

Unit tests were added to cover these cases.

Fixes bug #2685