Use collected static media for Django admin, and not the packaged media.
We were linking to the Django directory for the admin media files, but
these don't contain versioning information, causing failures on deployed
sites. Instead, we now need to use the collected media at package time.

The downside to this is that if a newer Django expects something in the
media, it may fail. We'll just have to keep updated and maybe require
very specific versions of Django, if this turns into a problem.