Begin moving to a new site style.
This is a refinement on the style we've had for years. The largish
page header and navigation bar have been slimmed down to a sleek two row
navigation bar at the top of the page, giving lots of room for the rest
of the page.

The search bar has moved to the top, beside the new account and support
dropdowns. The account dropdown links to preferences, log out, and the
administration UI (for admins). The support dropdown links to docs and
the bug reports.

There's also some style tweaks to the dashboard to tidy it up just a
little bit. These changes are more subtle.