Standardize the action bar color, and right-align it.
This removes the visual distinction between primary and secondary
actions. There's not much reason to really color these differently.
There's now a nice, seamless look across all actions.

The action bar has also been right-aligned, in order to match 1.6.x a
bit closer, so that muscle memory doesn't become an issue. In
particular, we want "View Diff" to be in the same location.