Properly serialize revisions in Diff upload API errors.
When a file couldn't be found in the repository during a diff upload
using the API, an error with the filename and revision would be
returned. If the revision was HEAD or PRE-CREATION, this would fail, due
to the JSON encoder being unable to serialize that sort of object.

Now we call unicode() on the revision to ensure it's always serialized
properly.