Fixed support for hg git diffs with spaces in path
Previously when HgDiffParser encountered a git format diff it parsed
the header by using spaces as a delimiter.  This falls apart when
passing in a diff to a file with spaces in it's path.

This commit updates the code to parse out the filenames using a regular
expression using the a/ and b/ prefixes of the before & after
filenames to mark the start of the filenames.

I've also updated the tests to check parsing of a git diff as generated
by the mercurial reviewboard extension.

Reviewed at http://reviews.reviewboard.org/r/3258/