Combine the Add/Update actions into an Update dropdown menu.
The Add/Update File/Screenshot/Diff menu items are now under an Update
menu. This is a bit nicer, and condenses the action bar a bit.

As part of this, I've removed the <img> for the dropdown arrow and made
the CSS take care of that. That required some shuffling of which part of
the items had background colors. It's now the <li>, not the <a>.