Fix several issues with the new site style.
The dropdown menus at the top of the page should no longer sit below
other UI elements. In fact, the whole z-order of elements is now defined
in one place, so that we're not using random numbers all over.

The borders of the review request action menus are now set correctly.

Primary column contents in the admin UI are no longer inheriting the
column header styling.

The trophies no longer cause the review request action bar to shift
down.