Introduce a file attachment image review UI.
This introduces the first real review UI for file attachments. There's a
few new JavaScript objects for easily dealing with
FileAttachmentComments and FileAttachment-based review UIs.

The UI is the same as what we have for screenshots. In fact, the
screenshot code has changed to just use the new image review UI code,
but with models for working with Screenshot and ScreenshotComments.

There's also some new code for registering review UIs, which will make
it easier to start developing new UIs. This includes in extensions.