Provide a model/view separation for comment blocks.
Comment blocks on a screenshot have been reworked into a model/view
setup, and are now designed to be reused outside of screenshot
commenting. This makes use of Backbone.js to define the model and view.

We now have an AbstractCommentBlock model that keeps track of serialized
comments coming from the page, as well as any draft comment being made,
and the total count of comments.

There's a ScreenshotCommentBlock that knows how to store coordinate
information and create ScreenshotComments.

And finally, a ScreenshotCommentBlockView that knows how to render this
to the screen.

The code is not ideal yet, largely because we're still working with
lots of older code, so we can't use such constructs as Backbone
collections. In time, these models will be cleaned up a bit.