Rewrite our screenshot review UI and introduce generic review infrastructure.
The screenshot review support has been rewritten to have a nice
model/view separation and to abstract much of the review process into
generic models and views.

There's now a few new classes: AbstractCommentBlock, AbstractCommentBlockView,
AbstractReviewable, and AbstractReviewableView. These handle the basics
of creating review UIs and comment blocks.

The screenshot subclasses, as a result, are fairly minimal for the most
part.

From here on out, it should be easier to create review UIs for custom
mimetypes. This will get even easier as more of our review
infrastructure is moved to new Backbone models and views.