Fix Git diffs with deleted binary files.
Deleted binary files in git diffs was causing parsing problems. We'd
skip over the entry entirely, so it would never show up in the diff
viewer, and it seems the file info was ending up in another file's diff.

The parsing has changed to not bail so early with empty (no-data)
changes. Really, this code could all use a rewrite, but this seems to
work pretty well for now.

Fixes bug #2419