Fixed the layout of widgets when the Request Statuses widget is open.
The Request Statuses widget builds a pie chart and legend dynamically.
The problem is, this happens after the widgets are layed out on the
page. The result is that the content ends up overlapping the widget
below.

There's now an event we can trigger after drawing that will tell the
admin UI to re-layout the widgets.