Fix a couple style issues with admin widgets headings.
Admin widget headings that were collapsed but had an action bar below it
would show a rounded header, which looked weird. Only the top half
should have been rounded. We now have a widget-has-actions class on the
widget to determine what to do here.

We also no longer indicate that the non-collapsable widgets (namely the
left-hand sidebar) are clickable.