Fix the display of headers in diff comments.
A call site wasn't converted to the new format for rendering in the
template, causing a problem when displaying headers for
functions/classes. This has been updated to return something the
template expects.

There was also an issue where an empty header was shown at times when no
headers were found. This has been made invisible again.