Fix webapi list of review requests when using ?shipit=0
When accessing the /api/review-requests resource with a shipit=0 parameter, it
would return the wrong results (often returning review requests that did have
reviews with ship-its). This code was initially written to do a join with the
reviews table and query the shipit parameter of that model, which was checking
whether *all* of the associated review requests didn't have ship-its (which
wasn't completely incorrect but not what a user really wants when passing
shipit=0).

To fix this, instead of excluding review requests with reviews that have ship
its, I've changed it to add a test for the shipit_count field on the
ReviewRequest model. I've also updated the unit tests for this API to use the
same query to test against.

While updating the unit tests, I discovered a couple issues. All of the
shipit_count fields in the fixtures were zero, despite there being the presence
of some reviews with a ship-it. I've updated the test_reviewrequests fixture to
fix this issue. I also had to add the test_site fixture to the
test_get_reviewrequests_with_ship_it_* tests in order for the data to be valid.

Fixes issue 2635.
Reviewed at http://reviews.reviewboard.org/r/3371/