Consolidate and fix the expand diff links.
There's now a couple template tags for rendering the diff links in a
simpler way. This ends up going through a single utility template to
render each link.

As part of this, we now have a smarter place to handle the diff headers.
We no longer show expansion links next to headers showing
functions/classes that are not contained within that collapsed region.
The headers are still shown, but you can't expand.