Fix incremental diff expansion on master.
Recent changes merged in for incremental diff expansion from
release-1.6.x needed some work.

First, the MEDIA_SERIAL/MEDIA_URL variables were removed from the diff
expand link context. This was needed for 1.6.x, but not for master, and
in fact won't work there. We no longer pass this in for the context to
that template.

Second, calls to expandChunk have been renamed to RB.expandChunk.