Document how to raise the MySQL packet size.
Very large diffs can cause errors with the mysql packet size. One request we've
gotten is to include documentation on how to fix this. This change does that,
but includes some pretty stern warnings that it's not a good idea in general.

This is based on Jesus Zambrano's change at
http://reviews.reviewboard.org/r/3364/