Redo the look and feel of the admin UI.
I wasn't happy with the new boxy look of the admin UI that was part of
the new widget work. They're fine for widgets, but were too prevelant
throughout the UI.

This changes the UI to have a sidebar dedicated to the permanent admin
UI panes (settings, common database actions), and to have a single box
on the right for content.

Some of the pages weren't following the proper page structure for
settings pages, and have been changed to fit in correctly.

Some visual issues with the buttons like "History" and "Add user" have
been fixed as well.

I'm planning to make further tweaks to this going forward, but this is
probably more or less what we'll ship with 1.7.0.