Fix version parsing for Mercurial.
Our version parsing was relying upon integers only in the version, and
had a probably incorrect hack for handling "+" in the version number.

Since Mercurial is a Python package following Python versioning rules,
we should be using parse_version comparisons. This is a lot less
fragile, and works with their new 1.4-rc release they just put out
(which completely breaks Review Board today).

This will be needed for an upcoming release soon.