Use the SHA1 password hasher for unit tests.
Django 1.4 moved to using PBKDF2 for handling password encryption. This
is much slower than the old default, SHA1, which can delay unit tests.

While we don't create many users in our unit tests, we seem to do this a
little bit. Forcing this to use SHA1 for unit tests speeds up the tests
a little bit, and keeps us sped up for future tests. It appears to save
about 6 seconds in our tests.