Add support for extra data in a FileAttachmentComment.
FileAttachmentComment now supports extra data, which is useful for
custom review UIs. These can be modified by sending any additional
fields in a POST or PUT request not normally handled by the resource.

The review UI support will soon make it easy to work with data in this
new field.