Add support for deleting user keys to SSHClient.
This allows SSHClient to delete a user key. It provides a
delete_user_key function on SSHClient and SSHStorage. This can be used
to provide a UI for deleting a user key.

This UI doesn't exist in Review Board yet, but will be easy to wire up
at some point.