Fix a couple path-related issues with rb-site and rbssh.
rb-site wasn't using the data directory as $HOME, and rbssh wasn't being
given the current PYTHONPATH. These weren't problems before, but became
issues with the new SSHDB key store.

We now set $HOME in rb-site to what we expect, and we pass sys.path as
PYTHONPATH when executing anything from an SCMTool.