Return 404 errors from the diff-comments resource.
When accessing the diff-comments resource, it would return a list of 0 items
instead of a 404 if the intermediate resource IDs were invalid. This change
fixes the URL to return 404 correctly, and also updates the FileDiffResource
queryset to use pk correctly in order to pass the unit tests.

Fixes bug 2552.
Reviewed at http://reviews.reviewboard.org/r/3401/