Allow automatic SSH key association when adding/editing GitHub repositories.
Defines an interface to allow a Review Board instance's SSH key to be
added to a hosting service's list of deploy keys for a given repository.
So far, only the GitHub HostingService implements the necessary methods
to upload the SSH key, and raises an exception when the upload fails.
The code is not currently being used by any module/app other than tests.

Reviewed at http://reviews.reviewboard.org/r/3423/